fix(phase8): restore ProductType RESERVATION, fix nearby 500, archive phase8 docs
- ProductType: GENERAL reverted to RESERVATION (예약 구매 / 선착순 구매 두 종류로 확정)
- V3 migration: fix product_type GENERAL → RESERVATION (DB 데이터 복원)
- ProductNearbyProjection: popularityScore Double → BigDecimal (MySQL DECIMAL 타입 불일치 수정)
- ProductMapper: popularityScore.toDouble() 변환 추가
- GlobalExceptionHandler: HttpMessageNotReadableException 400 핸들러 추가
- docs/pipeline-status.md: Phase 7/8 완료 + Phase 9~15 신규 로드맵 반영
- Phase 8 PDCA docs archived to docs/archive/2026-03/controller/

+++ domain-nearpick/src/main/kotlin/com/nearpick/nearpick/product/mapper/ProductMapper.kt
@@ -18,7 +18,7 @@ object ProductMapper {
         price = price,
         productType = ProductType.valueOf(productType),
         status = ProductStatus.valueOf(status),
-        popularityScore = popularityScore,
+        popularityScore = popularityScore.toDouble(),
         distanceKm = distanceKm,
         merchantName = merchantName,
         shopAddress = shopAddress,
